s3:printing: Fix size types
This fixes compilation with -Wstrict-overflow=2.

+++ source3/include/printing.h
@@ -63,7 +63,7 @@ enum {LPSTAT_OK, LPSTAT_STOPPED, LPSTAT_ERROR};
 
 typedef struct {
 	fstring message;
-	int qcount;
+	size_t qcount;
 	int status;
 }  print_status_struct;
